Test, Document, and improve performance of HelloMessageHandler.kt
Tests were written prior to refactoring, to ensure that functionality is not lost.

By utilizing the methods of the the DataOutputStream during message sending, we can avoid the extra allocations required when using a ByteBuffer and converting it to a ByteArray. By writing the Message directly to the stream, we can avoid additional allocations done by the Protocol Buffers library.

When parsing, we can request the Message to parse directly from the stream, which again, requires fewer allocations.